// src/api/__generated__/wallet/model/pageResponseUserSearchItemResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */
import type { UserSearchItemResponse } from './userSearchItemResponse';

/**
 * 普通分页响应
 */
export interface PageResponseUserSearchItemResponse {
  /** 当前页码 */
  page?: number;
  /** 每页数量 */
  pageSize?: number;
  /** 是否还有更多数据 */
  hasMore?: boolean;
  /** 分页数据 */
  list?: UserSearchItemResponse[];
}